+++ R/install.R
@@ -4,7 +4,7 @@ install_system <- function() {
       "Installing pre-commit into the conda environment",
       "`r-reticulate`."
     ))
-    install_precommit_impl()
+    install_impl()
     usethis::ui_done("Sucessfully installed pre-commit on your system.")
     usethis::ui_todo(
       "To use it with this project, run `precommit::use_precommit()`"
@@ -28,6 +28,28 @@ install_precommit <- function() {
   install_system()
 }
 
+#' Install pre-commit on your system with conda
+#' @keywords internal
+install_impl <- function() {
+  if (!"r-reticulate" %in% reticulate::conda_list()$name) {
+    reticulate::conda_create("r-reticulate")
+  }
+  reticulate::conda_install(packages = "pre-commit")
+}
+
+install_repo <- function(path_root) {
+  withr::with_dir(path_root, {
+    out <- call_and_capture(path_precommit_exec(), "install")
+    if (out$exit_status == 0) {
+      usethis::ui_done("Sucessfully installed pre-commit for repo {fs::path_file(path_root)}.")
+    } else {
+      usethis::ui_oops("Failed to install pre-commit for repo {fs::path_file(path_root)}.")
+      communicate_captured_call(out, preamble = "Problems during initialization:")
+    }
+  })
+}
+
+
 #' Unistall pre-commit
 #'
 #' Remove pre-commit from a repo or from your system.
@@ -47,19 +69,19 @@ uninstall_precommit <- function(scope = "repo",
   rlang::arg_match(ask, c("repo", "global", "both", "none"))
   withr::with_dir(path_root, {
     if (scope == "repo") {
-      uninstall_precommit_repo(ask = (ask %in% c("repo", "both")))
+      uninstall_repo(ask = (ask %in% c("repo", "both")))
       path_config <- ".pre-commit-config.yaml"
       if (fs::file_exists(path_config)) {
         fs::file_delete(path_config)
         usethis::ui_done("Removed .pre-commit-config.yaml")
       }
     } else if (scope == "global") {
-      uninstall_precommit_system(ask = (ask %in% c("global", "both")))
+      uninstall_system(ask = (ask %in% c("global", "both")))
     }
   })
 }
 
-uninstall_precommit_system <- function(ask = TRUE) {
+uninstall_system <- function(ask = TRUE) {
   if (is_installed()) {
     if (ask) {
       answer <- readline(paste(
@@ -87,12 +109,14 @@ uninstall_precommit_system <- function(ask = TRUE) {
           "Please remove pre-commit manually from the comamnd line. "
         ))
       } else {
-        out <- system2(
+        out <- call_and_capture(
           reticulate::conda_binary(),
           "remove -n r-reticulate pre-commit --yes"
         )
-        if (out == 0) {
+        if (out$exit_status == 0) {
           usethis::ui_done("Removed pre-commit from conda env r-reticulate.")
+        } else {
+          communicate_captured_call(out)
         }
       }
     } else {
@@ -106,7 +130,7 @@ uninstall_precommit_system <- function(ask = TRUE) {
   }
 }
 
-uninstall_precommit_repo <- function(ask) {
+uninstall_repo <- function(ask) {
   if (ask) {
     answer <- readline(paste0(
       "Are you sure you want to remove pre-commit from this repo? ",
@@ -119,12 +143,11 @@ uninstall_precommit_repo <- function(ask) {
     continue <- TRUE
   }
   if (continue) {
-    success <- grepl(
-      "pre-commit uninstalled",
-      system2(path_precommit_exec(), "uninstall", stdout = TRUE)
-    )
-    if (isTRUE(success)) {
+    out <- call_and_capture(path_precommit_exec(), "uninstall")
+    if (out$exit_status == 0) {
       usethis::ui_done("Uninstalled pre-commit from repo scope.")
+    } else {
+      communicate_captured_call(out)
     }
     if (is_package(".")) {
       lines <- readLines(".Rbuildignore", encoding = "UTF-8")
@@ -153,69 +176,6 @@ uninstall_precommit_repo <- function(ask) {
   } else {
     usethis::ui_info("You did not type 'yes', uninstallation process aborted.")
   }
-}
-
-#' Make a call with [system2()] and capture the effects.
-#' @param ... Arguments passed to [system2()].
-#' @return
-#' A list with:
-#' * content of stderr
-#' * content of stdout
-#' * exit status
-call_and_capture <- function(...) {
-  stdout <- tempfile()
-  writeLines("", stdout)
-  stderr <- tempfile()
-  writeLines("", stderr)
-  exit_status <- suppressWarnings(
-    system2(..., stdout = stdout, stderr = stderr)
-  )
-  stderr <- readLines(stderr)
-  if (exit_status != 0) {
-
-    if (length(stderr) < 1) {
-      stderr <- paste0(
-        "Could not recover stderr. Run the following command to get the error",
-        paste(...)
-      )
-    }
-  }
-  list(
-    stdout = readLines(stdout),
-    stderr = stderr,
-    exit_status = exit_status
-  )
-}
-
-install_repo <- function(path_root) {
-  tmp1 <- tempfile()
-  writeLines("", tmp1)
-  tmp2 <- tempfile()
-  writeLines("", tmp2)
-  withr::with_dir(path_root, {
-    out <- call_and_capture(path_precommit_exec(), "install")
-    if (out$exit_status == 0) {
-      usethis::ui_done("Sucessfully installed pre-commit for repo {fs::path_file(path_root)}.")
-    } else {
-      usethis::ui_oops("Failed to install pre-commit for repo {fs::path_file(path_root)}.")
-      communicate_captured_call(out, preamble = "Problems during initialization:")
-    }
-  })
-}
-
-#' @param x The output of [call_and_capture()].
-communicate_captured_call <- function(x, preamble = "") {
-  if (x$exit_status != 0) {
-    trans <- rlang::abort
-  } else {
-    trans <- rlang::warn
-  }
-  trans(paste0(preamble,
-    "\nstderr: ",
-    paste0(x$stderr, collapse = "\n"), "\n\nstdout: ",
-    paste0(x$stdout, collapse = "\n"),
-    collapse = "\n"
-  ))
 }
 
 is_installed <- function() {